README: List full command for convenience example
Since previously enough of the command was shown that it was easy to
miss the "..." in the middle, and accidentally copy the incomplete
command. This also switches to using 'echo ...', since it removes an
additional step for most people using the alias - and potentially avoids
problems if they omitted the quotes, or used double rather than single
quotes around the command.