README: Clarify path workarounds for Windows for both MSYS and MSYS2
It wasn't clear from the existing wording, that the additional leading
forward slash had to be added to both of the '-v' arguments, and not
just the first.

In addition, when using MSYS v2 or Cygwin, running PWD in 'C:\Users\Foo'
returns a path of '/home/Foo' rather than '/c/Users/Foo'. This won't
match the paths auto-created inside the Boot2Docker VM, so it's
necessary to use sed to adjust the path back to the expected format.