README: Set COMPOSE_PROJECT_NAME in the examples
By default Docker compose uses the current directory name as the project
name. The project name is used as a prefix for the image names. However
when using Compose via the dduportal/docker-compose image, the directory
name inside the container is always "app", which isn't helpful as a
project name. This can be overridden by setting COMPOSE_PROJECT_NAME in
the environment.